Improve performance of commit filters
